Polishing, cleanups. Some arrayCopy instead of doing that manually, so it should also be faster.

+++ src/de/lmu/ifi/dbs/elki/math/DoubleMinMax.java
@@ -102,7 +102,7 @@ public class DoubleMinMax extends DoubleDoublePair {
    */
   public void put(Collection<Double> data) {
     for(Double value : data) {
-      this.put(value);
+      this.put(value.doubleValue());
     }
   }
